chore(ct): rename component.d.ts
