+++ packages/playwright-ct-react17/hooks.d.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import type { JsonObject } from '@playwright/test/types/component';
+import type { JsonObject } from '@playwright/test/types/experimentalComponent';
 
 export declare function beforeMount<HooksConfig extends JsonObject>(
   callback: (params: { hooksConfig?: HooksConfig; App: () => JSX.Element }) => Promise<void | JSX.Element>
