+++ packages/playwright-ct-vue/hooks.d.ts
@@ -15,7 +15,7 @@
  */
 
 import type { App, ComponentPublicInstance } from 'vue';
-import type { JsonObject } from '@playwright/test/types/component';
+import type { JsonObject } from '@playwright/test/types/experimentalComponent';
 
 export declare function beforeMount<HooksConfig extends JsonObject>(
   callback: (params: { app: App; hooksConfig?: HooksConfig }) => Promise<void>
